BIG Changes
I think we need to handle the class imbalance problem more earlier.

This changes literally everything(correlations etc.). That's why I've created another copy. Difference between this and the original is, I've undersampled the dataset in the Data Preprocessing part in this file. We'll decide which one we should use next.